create loop state for reuse in the future.
